+++ source/databricks/settlement_report/settlement_report_job/entry_points/tasks/time_series_task.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from pyspark.sql import SparkSession
 
-from settlement_report_job.domain.time_series.order_by_columns import order_by_columns
+from settlement_report_job.domain.time_series_points.order_by_columns import order_by_columns
 from settlement_report_job.entry_points.tasks.task_type import TaskType
 from settlement_report_job.entry_points.tasks.task_base import TaskBase
 from settlement_report_job.infrastructure import csv_writer
@@ -11,7 +11,7 @@ from settlement_report_job.domain.utils.report_data_type import ReportDataType
 from settlement_report_job.entry_points.job_args.settlement_report_args import (
     SettlementReportArgs,
 )
-from settlement_report_job.domain.time_series.time_series_factory import (
+from settlement_report_job.domain.time_series_points.time_series_factory import (
     create_time_series_for_wholesale,
     create_time_series_for_balance_fixing,
 )
